fix: sidebar 구조 및 디자인 수정
