+++ src/App.vue
@@ -1,7 +1,4 @@
 <script setup>
-import { RouterLink, RouterView } from 'vue-router'
-
-import HelloWorld from './components/HelloWorld.vue'
 import HomeView from './views/HomeView.vue'
 import SideBarComponent from './components/sidebar/SideBarComponent.vue'
 </script>
