+++ src/components/sidebar/SideBarComponent.vue
@@ -10,22 +10,22 @@ const props = defineProps({
 
 const accountBooks = ref([
   {
-    id: "aasdfasdf",
+    id: "1",
     title: "KB마이핏통장"
   },
   {
-    id: "wersdfsdf",
+    id: "2",
     title: "KB모임통장 - 여행💛"
   }
-])
+]);
 </script>
 
 <template>
-  <div :class="['mt-3 sidebar', { show: props.isVisible }]">
-    <div class="mx-2 my-2 fw-bold">
+  <div :class="['sidebar', { show: props.isVisible }]">
+    <div class="mx-3 mt-3 mb-2 fw-bold">
       내 통장
     </div>
-    <button class="btn btn-account-book" type="button" v-for="accountBook in accountBooks">
+    <button class="btn btn-account-book" type="button" v-for="accountBook in accountBooks" :key="accountBook.id">
       💰 {{ accountBook.title }}
     </button>
   </div>
@@ -34,24 +34,29 @@ const accountBooks = ref([
 <style scoped>
 .sidebar {
   position: fixed;
-  top: 0;
-  left: 0;
-  width: 180px;
-  height: 80%;
-  background-color: #E8E1D7;
+  top: 60px;
+  left: -100px;
+  width: 200px;
+  height: 70%;
+  background-color: #F7F7F5;
+  box-shadow: 4px 7px 12px -1px gray;
   border-top-right-radius: 5%;
   border-bottom-right-radius: 5%;
+  transform: translateX(0);
   transform: translateX(-100%);
-  transition: transform 0.3s ease;
+  transition: transform 0.3s ease, opacity 0.3s ease;
+  opacity: 0;
+  z-index: 1100;
 }
 .sidebar.show {
+  left: 0;
   transform: translateX(0);
+  opacity: 1;
 }
 .btn-account-book {
   width: 100%;
   text-align: left;
   font-size: small;
-  /* background-color: saddlebrown; */
 }
 .btn-account-book:hover {
   background-color: #60584C;
